Moving existing log in page, register page and adding circular loading motion page
